Warnings
git-svn-id: file:///Users/tack/GecodeGitMigration/gecode-svn-mirror/gecode/trunk@11192 e85b7adc-8362-4630-8c63-7469d557c915

+++ gecode/int/array.hpp
@@ -66,8 +66,8 @@ namespace Gecode {
   }
   forceinline
   IntArgs::IntArgs(const std::vector<int>& x)
-    : PrimArgArray<int>(x.size()) {
-    for (int i=x.size(); i--;)
+    : PrimArgArray<int>(static_cast<int>(x.size())) {
+    for (unsigned int i=x.size(); i--;)
       a[i] = x[i];
   }
